// src/types/user.ts
export interface User{
    id: string ;
    name: string;
    email: string;
    identification_number: string;
    address: string;
    phone: string;
    password: string;
    profile_picture: string;
    created_at: string;
    updated_at: string;
    work_info_id: string | null;
}